Add support for passing a database URL and ssl config
When running on with [PostgreSQL on Heroku](https://devcenter.heroku.com/articles/heroku-postgresql#connecting-in-node-js), the connection parameters are provided as a URL string. The pg package supports this, it's just a matter of passing it through the pdf-bot config.

Additionally, this PR adds support for the other possible `pg` [client options](https://node-postgres.com/api/client).